Migrate token-amount-input
