+++ packages/harmony/src/components/input/index.ts
@@ -2,3 +2,4 @@ export * from './SelectablePill'
 export * from './TextInput'
 export * from './useFocusState'
 export * from './PasswordInput'
+export * from './TokenAmountInput'
