+++ packages/mobile/src/harmony-native/components/index.ts
@@ -8,6 +8,7 @@ export * from './input/TextInput/TextInput'
 export * from './input/TextInput/types'
 export * from './input/PasswordInput/PasswordInput'
 export * from './input/PasswordInput/types'
+export * from './input/TokenAmountInput'
 export * from './CoverPhoto/CoverPhoto'
 export * from './Button/Button/Button'
 export * from './Button/Button/types'
