--- packages/stems/src/components/TokenAmountInput/TokenAmountInput.module.css
@@ -1,50 +1,0 @@
-.root {
-  width: 100%;
-}
-
-.inputWrapper {
-  display: flex;
-  align-items: center;
-  width: 100%;
-  box-sizing: border-box;
-  background: var(--neutral-light-10);
-  border: 1px solid var(--neutral-light-8);
-  border-radius: 6px;
-}
-.input {
-  flex: 1;
-  padding: 16px;
-  border: 0;
-  outline: 0;
-  background: 0;
-  color: var(--neutral);
-  font-size: var(--font-xl);
-  font-weight: var(--font-medium);
-}
-.input:focus {
-  outline: none;
-}
-.input::placeholder {
-  /* Chrome, Firefox, Opera, Safari 10.1+ */
-  color: var(--neutral-light-4);
-  mix-blend-mode: normal;
-  opacity: 0.5;
-}
-
-.tokenLabel {
-  padding-right: 16px;
-  color: var(--neutral);
-  font-size: var(--font-xl);
-  font-weight: var(--font-bold);
-}
-
-.label {
-  display: block;
-  margin-bottom: 12px;
-  font-family: var(--font-family);
-  font-weight: var(--font-bold);
-  font-size: var(--font-s);
-  letter-spacing: 1px;
-  text-transform: uppercase;
-  color: var(--neutral-light-4);
-}
