--- packages/stems/src/components/TokenValueInput/TokenValueInput.module.css
@@ -1,49 +1,0 @@
-.container {
-  position: relative;
-}
-
-.label {
-  font-family: var(--font-family);
-  font-weight: var(--font-bold);
-  text-align: left;
-  font-size: var(--font-s);
-  line-height: 17px;
-  letter-spacing: 1px;
-  text-transform: uppercase;
-  color: var(--neutral-light-4);
-  margin-bottom: 11px;
-}
-
-.input {
-  padding: 14px 16px;
-  background: rgba(0, 0, 0, 0.05);
-  border: 1px solid var(--neutral-light-8);
-  border-radius: 6px;
-  color: var(--neutral);
-  mix-blend-mode: normal;
-  width: 100%;
-  box-sizing: border-box;
-}
-
-.input::placeholder {
-  /* Chrome, Firefox, Opera, Safari 10.1+ */
-  color: var(--neutral-light-4);
-  mix-blend-mode: normal;
-  opacity: 0.5;
-}
-
-.input:focus {
-  outline: none;
-}
-
-.rightLabel {
-  position: absolute;
-  bottom: 15px;
-  right: 10px;
-  color: var(--neutral-light-4);
-}
-
-textarea {
-  resize: none;
-  height: 100%;
-}
