--- packages/stems/src/components/TokenValueInput/index.ts
@@ -1,2 +1,0 @@
-export { TokenValueInput } from './TokenValueInput'
-export { TokenValueInputProps, Format } from './types'
