+++ packages/stems/src/index.ts
@@ -62,16 +62,6 @@ export {
   TokenValueSlider,
   TokenValueSliderProps
 } from './components/TokenValueSlider'
-export {
-  TokenValueInput,
-  TokenValueInputProps,
-  Format
-} from './components/TokenValueInput'
-export {
-  TokenAmountInput,
-  TokenAmountInputProps,
-  TokenAmountInputChangeHandler
-} from './components/TokenAmountInput'
 
 export { useHotkeys } from './hooks/useHotKeys'
 export { useClickOutside } from './hooks/useClickOutside'
