+++ packages/web/src/pages/audio-rewards-page/components/SendInputBody.module.css
@@ -1,7 +1,3 @@
-.titleContainer {
-  margin: 24px auto;
-}
-
 .subtitle {
   font-weight: 500;
   font-size: 16px;
@@ -54,11 +50,6 @@
   line-height: 20px;
   border: none;
   color: var(--neutral);
-}
-
-/* Button */
-.sendBtn {
-  margin: 24px;
 }
 
 .sendBtnText {
